REFACTOR : context와 타입 분리
